test: update profile information form test
